F #2830: GOCA - host split and add state.go file
